+++ src/kpir/app/products/model.py
@@ -8,5 +8,5 @@ class Products(db.Model, CoreModel):
     name = db.Column(db.String(255), index=True, nullable=False)
     sku = db.Column(db.String(64), unique=True)
     price = db.Column(db.Float, nullable=False)
-    tax = db.Column(db.Integer(3), nullable=False)
+    tax = db.Column(db.Integer, nullable=False)
     jm = db.Column(db.String(24))
